Revert "Refactor: put interfaces into subfolder interfaces"
This reverts commit 00b823ce0c76a9f944129783dce07a91a709601e.
