+++ src/ai/recommendation/pickMostRecommended.ts
@@ -1,4 +1,4 @@
-import { IWallpaper } from '../../utils/interfaces/IWallpaper';
+import { IWallpaper } from '../../utils/IWallpaper';
 import { number_likeness } from '../../utils/typeAliases';
 import { IWallpaperVector } from './IWallpaperVector';
 import { wallpaperToVector } from './wallpaperToVector';
