+++ src/ai/recommendation/wallpaperToVector.ts
@@ -1,4 +1,4 @@
-import { IWallpaper } from '../../utils/interfaces/IWallpaper';
+import { IWallpaper } from '../../utils/IWallpaper';
 import { IWallpaperVector } from './IWallpaperVector';
 
 export function wallpaperToVector(wallpaper: IWallpaper): IWallpaperVector {
