+++ src/components/Gallery/GalleryFilter/utils/parseKeywordsFromWallpaper.ts
@@ -1,7 +1,7 @@
 import { IKeywords, parseKeywords } from 'n12';
-import { IWallpaper } from '../../../../utils/interfaces/IWallpaper';
+import { IWallpaper } from '../../../../utils/IWallpaper';
 
-export function parseKeywordsFromWallpaper(wallpaper: Pick<IWallpaper, 'prompt' | 'content'>): IKeywords {
+export function parseKeywordsFromWallpaper(wallpaper: Pick<IWallpaper,'prompt'|'content'>): IKeywords {
     const { prompt, content } = wallpaper;
     return parseKeywords({ prompt, content });
 }
