+++ src/export/getWallpaperBaseFilename.ts
@@ -1,10 +1,10 @@
 import { normalizeToKebabCase } from 'n12';
-import { IWallpaper } from '../utils/interfaces/IWallpaper';
+import { IWallpaper } from '../utils/IWallpaper';
 
 /**
  * Function that returns the base filename of a wallpaper
- *
- *
+ * 
+ * 
  * @param wallpaper - The wallpaper object.
  * @returns The base filename of the wallpaper.
  */
