+++ src/utils/IMidjourneyJob.ts
