+++ src/utils/hooks/WallpapersContext.ts
@@ -1,5 +1,5 @@
 import { createContext } from 'react';
 import { BehaviorSubject } from 'rxjs';
-import { IWallpaper } from '../interfaces/IWallpaper';
+import { IWallpaper } from '../IWallpaper';
 
 export const WallpapersContext = createContext<Record<string /*_uuid*/, BehaviorSubject<IWallpaper>>>({});
